refactor: *content controller

+++ src/main/java/org/omega/omegapoisk/content/service/AnimeContentService.java
@@ -24,7 +24,7 @@ public class AnimeContentService extends AbstractContentService<Anime> {
     private final AvgRatingService avgRatingService;
 
     @Value("${spring.application.page}")
-    private int page;
+    private long page;
 
     public AnimeContentService(AnimeRepository animeRepository, AvgRatingService avgRatingService) {
         super(animeRepository, avgRatingService);
